Move timestamp methods into nss_cache/util/timestamps.py.
These will be reused in the GCS source, since the GCS API
returns datetime objects like the S3 API.
