+++ nss_cache/util/timestamps.py
@@ -22,6 +22,7 @@ import os.path
 import tempfile
 import time
 import stat
+import datetime
 
 
 def ReadTimestamp(filename):
@@ -119,3 +120,26 @@ def WriteTimestamp(timestamp, filename):
     os.rename(temp_filename, filename)
     logging.debug('wrote timestamp %s to file %r', time_string, filename)
     return True
+
+
+def FromTimestampToDateTime(ts):
+    """Converts internal nss_cache timestamp to datetime object.
+
+    Args:
+      ts: number of seconds since epoch
+    Returns:
+      datetime object
+    """
+    return datetime.datetime.utcfromtimestamp(ts)
+
+
+def FromDateTimeToTimestamp(datetime_obj):
+    """Converts datetime object to internal nss_cache timestamp.
+
+    Args:
+      datetime object
+    Returns:
+      number of seconds since epoch
+    """
+    dt = datetime_obj.replace(tzinfo=None)
+    return int((dt - datetime.datetime(1970, 1, 1)).total_seconds())
